+++ TechMeter.API/Controllers/LessonController.cs
@@ -9,16 +9,12 @@ using System.Security.Claims;
 using TechMeter.Application.DTO.Lesson;
 using TechMeter.Application.DTO.LessonComment;
 using TechMeter.Application.Features.Lesson.Command.AddLesson;
-using TechMeter.Application.Features.Lesson.Command.AddLessonComment;
 using TechMeter.Application.Features.Lesson.Command.ChangeLessonState;
 using TechMeter.Application.Features.Lesson.Command.DeleteLesson;
-using TechMeter.Application.Features.Lesson.Command.DeleteLessonComment;
 using TechMeter.Application.Features.Lesson.Command.EditLesson;
-using TechMeter.Application.Features.Lesson.Command.EditLessonComment;
 using TechMeter.Application.Features.Lesson.Command.UnWatchLesson;
 using TechMeter.Application.Features.Lesson.Query.GetAllLessons;
 using TechMeter.Application.Features.Lesson.Query.GetLessonById;
-using TechMeter.Application.Features.Lesson.Query.GetLessonComments;
 using TechMeter.Application.Features.Lesson.Query.GetSectionLessons;
 using TechMeter.Application.Features.Lesson.Query.StudentLessonWatched;
 //using TechMeter.Application.Interfaces.Lesson;
