--- TechMeter.Application/Features/Lesson/Command/UnLikeOnLessonComment/UnLikeOnLessonCommentCommand.cs
@@ -1,13 +1,0 @@
-﻿using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TechMeter.Domain.Shared.Bases;
-
-namespace TechMeter.Application.Features.Lesson.Command.LikeOnLessonComment
-{
-    public sealed record UnLikeOnLessonCommentCommand(string commentId, string userId) : IRequest<Response<string>>;
-
-}
