+++ TechMeter.Application/Features/Payment/Command/PaymentIntent/PaymentIntentCommandHandler.cs
@@ -5,7 +5,6 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using TechMeter.Application.DTO.Payment;
-//using TechMeter.Application.Interfaces.Payment;
 using TechMeter.Application.Interfaces.Services.Payment;
 using TechMeter.Domain.Shared.Bases;
